Precompile advection diffusion folder
